# Use while loop instead of for-range loop in broadcast kernel